FIx bug fix float integer

+++ src/schemes/response/user_order.go
@@ -16,7 +16,7 @@ type UserOrder struct {
 type OrderItem struct {
 	ID     uuid.UUID `json:"id" gorm:"column:dish_id"`
 	Image  *string   `json:"image" gorm:"column:path"`
-	Price  int       `json:"price"`
+	Price  float32       `json:"price"`
 	Name   string    `json:"name"`
 	Amount int       `json:"amount"`
 }
